Fix: Move and rename classes
